requirements.txt
# Requires Python >= 3.10 (uses PEP 604 union syntax: int | None)
pandas~=2.2.3
pyodbc~=5.2.0
openpyxl~=3.1.5
python-dotenv~=1.2.1
numpy~=2.2.6
pyarrow~=18.1.0
pywin32~=310; sys_platform == "win32"
fpdf2~=2.8.6
Pillow~=11.1.0
